+++ src/NServiceBus.Storage.MongoDB.Tests/Sagas/Migration/When_migrating_NServiceBus_dot_MongoDB.cs
@@ -2,29 +2,37 @@
 {
     using System;
     using System.Threading.Tasks;
+    using global::MongoDB.Bson;
     using NUnit.Framework;
     using Persistence.ComponentTests;
 
-    class When_migrating_NServiceBus_dot_Persistence_dot_MongoDb : SagaMigrationPersisterTests
+    class When_migrating_NServiceBus_dot_MongoDB : SagaMigrationPersisterTests
     {
         [Test]
         public async Task Persister_works_with_existing_sagas()
         {
-            var legacySagaData = new Persistence.MongoDB.NServiceBusPersistenceMongoDBLegacySagaData
+            var legacySagaData = new NServiceBus.MongoDB.NServiceBusMongoDBLegacySagaData
             {
                 Id = Guid.NewGuid(),
                 OriginalMessageId = Guid.NewGuid().ToString(),
                 Originator = "SomeOriginator",
-                Version = 5,
+                DocumentVersion = 5,
                 SomeCorrelationPropertyId = Guid.NewGuid(),
                 SomeUpdatableSagaData = GetHashCode()
             };
 
-            configuration = new PersistenceTestsConfiguration(nameof(legacySagaData.Version));
+            configuration = new PersistenceTestsConfiguration(nameof(legacySagaData.DocumentVersion), collectionNamingConvention);
 
-            await PrepareSagaCollection(legacySagaData, nameof(legacySagaData.SomeCorrelationPropertyId));
+            await PrepareSagaCollection(legacySagaData, nameof(legacySagaData.SomeCorrelationPropertyId), d =>
+            {
+                var document = d.ToBsonDocument();
 
-            var retrievedSagaData = await GetById<NServiceBusPersistenceMongoDBLegacySaga, NServiceBusPersistenceMongoDBLegacySagaData>(legacySagaData.Id);
+                d.ETag = document.GetHashCode();
+
+                return d.ToBsonDocument();
+            });
+
+            var retrievedSagaData = await GetById<NServiceBusMongoDBLegacySagaData>(legacySagaData.Id);
 
             Assert.IsNotNull(retrievedSagaData, "Saga was not retrieved");
             Assert.AreEqual(legacySagaData.OriginalMessageId, retrievedSagaData.OriginalMessageId, "OriginalMessageId does not match");
@@ -33,20 +41,22 @@
             Assert.AreEqual(legacySagaData.SomeUpdatableSagaData, retrievedSagaData.SomeUpdatableSagaData, "SomeUpdatableSagaData does not match");
         }
 
-        class NServiceBusPersistenceMongoDBLegacySaga : Saga<NServiceBusPersistenceMongoDBLegacySagaData>, IAmStartedByMessages<MigrationStartMessage>
+        readonly Func<Type, string> collectionNamingConvention = t => t.Name;
+
+        class NServiceBusMongoDBLegacySaga : Saga<NServiceBusMongoDBLegacySagaData>, IAmStartedByMessages<MigrationStartMessage>
         {
             public Task Handle(MigrationStartMessage message, IMessageHandlerContext context)
             {
                 throw new NotImplementedException();
             }
 
-            protected override void ConfigureHowToFindSaga(SagaPropertyMapper<NServiceBusPersistenceMongoDBLegacySagaData> mapper)
+            protected override void ConfigureHowToFindSaga(SagaPropertyMapper<NServiceBusMongoDBLegacySagaData> mapper)
             {
                 mapper.ConfigureMapping<MigrationStartMessage>(msg => msg.Id).ToSaga(saga => saga.Id);
             }
         }
 
-        class NServiceBusPersistenceMongoDBLegacySagaData : IContainSagaData
+        class NServiceBusMongoDBLegacySagaData : IContainSagaData
         {
             public Guid SomeCorrelationPropertyId { get; set; }
 
@@ -60,13 +70,15 @@
     }
 }
 
-namespace NServiceBus.Persistence.MongoDB
+namespace NServiceBus.MongoDB
 {
     using System;
 
-    class NServiceBusPersistenceMongoDBLegacySagaData : IContainSagaData
+    class NServiceBusMongoDBLegacySagaData : IContainSagaData
     {
-        public int Version { get; set; }
+        public int DocumentVersion { get; set; } //From NServiceBus.MongoDB.IHaveDocumentVersion
+
+        public int ETag { get; set; } //From NServiceBus.MongoDB.IHaveDocumentVersion
 
         public Guid SomeCorrelationPropertyId { get; set; }
