+++ src/NServiceBus.Storage.MongoDB.Tests/Sagas/Migration/When_migrating_NServiceBus_dot_Persistence_dot_MongoDb.cs
@@ -2,37 +2,29 @@
 {
     using System;
     using System.Threading.Tasks;
-    using global::MongoDB.Bson;
     using NUnit.Framework;
     using Persistence.ComponentTests;
 
-    class When_migrating_NServiceBus_dot_MongoDB : SagaMigrationPersisterTests
+    class When_migrating_NServiceBus_dot_Persistence_dot_MongoDb : SagaMigrationPersisterTests
     {
         [Test]
         public async Task Persister_works_with_existing_sagas()
         {
-            var legacySagaData = new NServiceBus.MongoDB.NServiceBusMongoDBLegacySagaData
+            var legacySagaData = new Persistence.MongoDB.NServiceBusPersistenceMongoDBLegacySagaData
             {
                 Id = Guid.NewGuid(),
                 OriginalMessageId = Guid.NewGuid().ToString(),
                 Originator = "SomeOriginator",
-                DocumentVersion = 5,
+                Version = 5,
                 SomeCorrelationPropertyId = Guid.NewGuid(),
                 SomeUpdatableSagaData = GetHashCode()
             };
 
-            configuration = new PersistenceTestsConfiguration(nameof(legacySagaData.DocumentVersion), collectionNamingConvention);
+            configuration = new PersistenceTestsConfiguration(nameof(legacySagaData.Version));
 
-            await PrepareSagaCollection(legacySagaData, nameof(legacySagaData.SomeCorrelationPropertyId), d =>
-            {
-                var document = d.ToBsonDocument();
+            await PrepareSagaCollection(legacySagaData, nameof(legacySagaData.SomeCorrelationPropertyId));
 
-                d.ETag = document.GetHashCode();
-
-                return d.ToBsonDocument();
-            });
-
-            var retrievedSagaData = await GetById<NServiceBusMongoDBLegacySaga, NServiceBusMongoDBLegacySagaData>(legacySagaData.Id);
+            var retrievedSagaData = await GetById<NServiceBusPersistenceMongoDBLegacySagaData>(legacySagaData.Id);
 
             Assert.IsNotNull(retrievedSagaData, "Saga was not retrieved");
             Assert.AreEqual(legacySagaData.OriginalMessageId, retrievedSagaData.OriginalMessageId, "OriginalMessageId does not match");
@@ -41,22 +33,20 @@
             Assert.AreEqual(legacySagaData.SomeUpdatableSagaData, retrievedSagaData.SomeUpdatableSagaData, "SomeUpdatableSagaData does not match");
         }
 
-        readonly Func<Type, string> collectionNamingConvention = t => t.Name;
-
-        class NServiceBusMongoDBLegacySaga : Saga<NServiceBusMongoDBLegacySagaData>, IAmStartedByMessages<MigrationStartMessage>
+        class NServiceBusPersistenceMongoDBLegacySaga : Saga<NServiceBusPersistenceMongoDBLegacySagaData>, IAmStartedByMessages<MigrationStartMessage>
         {
             public Task Handle(MigrationStartMessage message, IMessageHandlerContext context)
             {
                 throw new NotImplementedException();
             }
 
-            protected override void ConfigureHowToFindSaga(SagaPropertyMapper<NServiceBusMongoDBLegacySagaData> mapper)
+            protected override void ConfigureHowToFindSaga(SagaPropertyMapper<NServiceBusPersistenceMongoDBLegacySagaData> mapper)
             {
                 mapper.ConfigureMapping<MigrationStartMessage>(msg => msg.Id).ToSaga(saga => saga.Id);
             }
         }
 
-        class NServiceBusMongoDBLegacySagaData : IContainSagaData
+        class NServiceBusPersistenceMongoDBLegacySagaData : IContainSagaData
         {
             public Guid SomeCorrelationPropertyId { get; set; }
 
@@ -70,15 +60,13 @@
     }
 }
 
-namespace NServiceBus.MongoDB
+namespace NServiceBus.Persistence.MongoDB
 {
     using System;
 
-    class NServiceBusMongoDBLegacySagaData : IContainSagaData
+    class NServiceBusPersistenceMongoDBLegacySagaData : IContainSagaData
     {
-        public int DocumentVersion { get; set; } //From NServiceBus.MongoDB.IHaveDocumentVersion
-
-        public int ETag { get; set; } //From NServiceBus.MongoDB.IHaveDocumentVersion
+        public int Version { get; set; }
 
         public Guid SomeCorrelationPropertyId { get; set; }
